Enhance API Key Management in Dashboard
